adding service check script

+++ installers/setup-terraform.sh
@@ -1,17 +1,29 @@
-#!/bin/sh
+#!/bin/bash
+# Author: Abhishek Anand Amralkar
+# This script installs Terraform
 
-# Fail on error
-set -e
+unset CDPATH
+CURDIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
+TF_VERSION=${TF_VERSION:-"0.11.12"}
+TF_BIN=${TF_BIN:-"~/bin/terraform"}
 
-# Update deps
-sudo apt-get update -y 
-sudo apt-get upgrade -y
+install_terraform (){
+    if [ ! -e "$TF_BIN" ];
+     then
+           echo "Installing Terraform..."
+           wget --no-check-certificate  https://releases.hashicorp.com/terraform/${TERRAFORM_VERSION}/terraform_${TERRAFORM_VERSION}_linux_amd64.zip 
+           sudo unzip terraform_${TERRAFORM_VERSION}_linux_amd64.zip
+           sudo mv terraform /home/aaa/bin/
+           echo "Done!"
+    else
+        
 
-TERRAFORM_VERSION="0.11.12"
 
-echo "Installing Terraform..."
-wget --no-check-certificate  https://releases.hashicorp.com/terraform/${TERRAFORM_VERSION}/terraform_${TERRAFORM_VERSION}_linux_amd64.zip 
-sudo unzip terraform_${TERRAFORM_VERSION}_linux_amd64.zip
-sudo mv terraform /home/aaa/bin/
-
-echo "Done!"
+install_leingen () {
+    if [ ! -e "$LEIN_BIN" ];
+    then
+      cd /tmp && curl -O https://raw.githubusercontent.com/technomancy/leiningen/stable/bin/lein && sudo chmod 755 lein && sudo mv lein /usr/bin && sudo chmod a+x ~/bin/lein
+    else
+      echo "Lein is installed"
+      fi
+}
